всё сделано по примеру с github, но не работает

+++ sources/Device/src/Hardware/I2C/i2c.h
@@ -18,7 +18,7 @@ int8_t user_i2c_read(uint8_t dev_id, uint8_t reg_addr, uint8_t *reg_data, uint16
 int8_t user_i2c_write(uint8_t dev_id, uint8_t reg_addr, uint8_t *reg_data, uint16_t len);
 
 int8_t user_i2c_write8(uint8_t dev_id, uint8_t data);
-int8_t user_i2c_read16(uint8_t dev_id, uint16_t *data);
+int8_t user_i2c_read16(uint8_t dev_id, uint8_t *data);
 #ifdef __cplusplus
 }
 #endif
